changed month from long to short

+++ src/utils/functions.ts
@@ -8,7 +8,7 @@ const formatNumberWithCommas = (number: number | undefined): string => {
 const displayMonthAndYear = (dateString: Date | undefined): string =>{
     if(dateString){
         const date = new Date(dateString);
-       return  `${date.toLocaleString('en', { month: 'long' })} ${date.getFullYear()}`;
+       return  `${date.toLocaleString('en', { month: 'short' })} ${date.getFullYear()}`;
     }
 
     return "";
